refactor: move paradigm generation test to its own file
